adapt for Gnome3
Under Gnome3 SCREEN_SIZE is set to maximum (there is no need to provide space for panel which
contains button for showing the graphics window, because there is no panel with window
selection buttons in Gnome3).

Without this change Gnome3 will occupy the screen space with the useless upper panel which shows
time.